Change folder structure for config
