+++ gatsby/onCreatePage.js
@@ -1,0 +1,79 @@
+const _ = require('lodash');
+const path = require('path');
+
+/**
+ * Create all blog post pages
+ *
+ * @param {*} { boundActionCreators, graphql }
+ * @returns
+ */
+module.exports = ({ boundActionCreators, graphql }) => {
+  const { createPage } = boundActionCreators;
+
+  // Query data used for posts
+  return graphql(`
+    {
+      allMarkdownRemark(limit: 1000) {
+        edges {
+          node {
+            id
+            fields {
+              slug
+            }
+            frontmatter {
+              tags
+              templateKey
+            }
+          }
+        }
+      }
+    }
+  `).then(result => {
+    if (result.errors) {
+      result.errors.forEach(e => console.error(e.toString()));
+      return Promise.reject(result.errors);
+    }
+
+    const posts = result.data.allMarkdownRemark.edges;
+
+    // Loop every markdown file and create a post page for each one
+    posts.forEach(edge => {
+      const id = edge.node.id;
+      createPage({
+        path: edge.node.fields.slug,
+        tags: edge.node.frontmatter.tags,
+        component: path.resolve(
+          `src/templates/${String(edge.node.frontmatter.templateKey)}.js`,
+        ),
+        // Pass additional data
+        context: {
+          id,
+        },
+      });
+    });
+
+    let tags = [];
+    // Loop every post and put all found tags into `tags`
+    posts.forEach(edge => {
+      if (_.get(edge, `node.frontmatter.tags`)) {
+        tags = tags.concat(edge.node.frontmatter.tags);
+      }
+    });
+
+    // Eliminate duplicate tags
+    tags = _.uniq(tags);
+
+    // Create a tag page for each tag
+    tags.forEach(tag => {
+      const tagPath = `/tags/${_.kebabCase(tag)}/`;
+
+      createPage({
+        path: tagPath,
+        component: path.resolve(`src/templates/tags.js`),
+        context: {
+          tag,
+        },
+      });
+    });
+  });
+};
